Dont parse quote twice
